+++ ChangeLog.md
@@ -10,6 +10,7 @@ More detailed information about incremental changes can be found in the
 ### 3.14.1 (2021-xx-yy)
 
 - Fixed invalidation of cached Hessians when reoptimizing with same structure.
+- Added `OptionsList::SetBoolValue()` and `OptionsList::SetBoolValueIfUnset()`. [#492]
 
 ### 3.14.0 (2021-06-15)